completed styling of header
